Multilane basics

+++ src/helpers.h
@@ -14,4 +14,4 @@ int find_next(std::vector<Car *> &lane, int idx);
 // return distance to previous car
 int find_prev(std::vector<Car *> &lane, int idx);
 
-int weighted_random_choice(std::map<int, float> &dict);
+int weighted_random_choice(std::map<int, int> &dict);
